Move classes, use MaintenanceModuleInterface and rewrite request token
