+++ contao/templates/backend/be_theme_compiler.html5
@@ -24,7 +24,7 @@
         <form action="<?= $this->action ?>" class="tl_form" method="get">
             <input type="hidden" name="act" value="compile">
             <input type="hidden" name="do" value="maintenance">
-            <input type="hidden" name="rt" value="<?= REQUEST_TOKEN ?>">
+            <input type="hidden" name="rt" value="<?= $this->requestToken ?>">
             <fieldset class="tl_tbox nolegend">
                 <div class="w50">
                     <h3><label for="ctrl_theme">Theme</label></h3>
