+++ src/DependencyInjection/ContaoThemeCompilerExtension.php
@@ -20,11 +20,8 @@ class ContaoThemeCompilerExtension extends Extension
 {
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new YamlFileLoader(
-            $container,
-            new FileLocator(__DIR__.'/../../config')
-        );
-
-        $loader->load('commands.yml');
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../../config'));
+        $loader->load('services.yaml');
+        $loader->load('commands.yaml');
     }
 }
